Create SECURITY.markdown file
Added SECURITY.markdown file through the Github web-interface.  Copied the text of SECURITY_DISCLOSURE.markdown in to this file.  There does not appear to be a way to tell Github that we already have a security disclosure process defined in a different file to the one it expects in its web interface.