Integrated the diesel migrations into the PAS itself
This avoids needing to run populate-test-database.sh to create
an initial database. However, this raises a question of if it's
correct behavior for the PAS to include some fixed hash of the root
enclaves.

I'm currently wondering if a better alternative would be to provide the
binaries/hash to the PAS during setup.

Need to revisit.