Added renamable paths to pgen's --binary argument
This allows file name remapping similar to what is in veracruz-client at
the time of writing this (may be changed later).

pgen --binary example-binary.wasm:example/example-binary.wasm

Without this it's not possible to use pgen in a directory outside of
where the binary lives, since the directory component is mapped into the
enclave's filesystem, resulting in an invalid path.

I believe only the --binary option in pgen needs since, since it is the
only option used for both the enclave's filesystem and the user's local
filesystem (to indicate the binary name and to open the binary for
hashing).