Remove empty directory ("") from the VFS to avoid conflicts with the root directory ("/").
In WASI, directories are discovered (preopened) at startup, starting from file descriptor fd=3 and incrementing fd until until failure (cf. `__wasilibc_populate_preopens()` in wasi-libc).
An internal register indexes those file descriptors associated with their path.
This register is checked before calling `path_open()`, which requires the parent directory's file descriptor of the path to open.
Because any forward flash prefixing the path is stripped away when a directory is saved to the registry, the root directory ("/") ends up with the same prefix as the empty directory ("").
And since the most recent entries take precedence over the least recent ones, the result of determining the parent directory of a file under the root directory always defaults to the empty directory, which is not currently supported in our implementation of `path_open()`