Move some stuff from fs.rs into common.rs.
In preparation for implementing an strace-like function,
move some stuff that does not appear to be file-system-related
from fs.rs into common.rs. This includes the configuration
parameter enable_clock, which is now part of a struct Options
that can easily be extended. The main system calls affected are:

* clock_res_get
* clock_time_get
* random_get

The behaviour of fd_filestat_set_times and path_filestat_set_times
is slightly altered: they will no longer fail if the clock is not
enabled.