Added automatic translation of relative paths to absolute paths in CLIs
For example:

    vc-client --data my-data=path/to/local/data

becomes

    vc-client --data /my-data=path-to/local/data

This makes the adoption of directory support trivial, and should reduce
simple mistakes from forgetting the leading slash. If Veracruz had a
concept of current working directory (cwd), this would be the same
behavior as if the cwd is always the root. This is common on systems
without cwd state.

Note there's an open question which layer should implicitly translate to
absolute paths. I've done the translation here in veracruz_client, which
seems to be where most pure-qol features happen. This means in both the
transport_protocol and policy files, absolute paths are still required.
Relative paths follow the current behavior, which is to error with EACCES.