Added option to reference computation by policy hash, changed to use query strings
Per feedback from @alexandref75

It was pointed out for most cases 1. the policy hash uniquely identifies
the computation, and 2. the hash is easily available and can be baked
into applications.

The currently lookup could be slow if we have many computation
instances, a second HashMap with hashes as keys would be more
performant, but this gets messy with Rust's ownership system.