icecap: docker: Introduce stateful configuration
The original Docker configuration for hacking on IceCap uses a
persistent `/nix` directory, which is bind-mounted into the container.
This makes obtaining a clean Docker environment much simpler. However,
bind mounts are unsuitable for this kind of use in the case of Docker on
Mac. Not only are they very slow, but this kind of filesystem access
results odd states that cause crashes.

This commit adds a so-called stateful configuration, where the `/nix`
directory is located inside the container. This configuration is more
robust on MacOS, at the cost of making obtaining a clean development
environment more expensive.

This new stateful configuration is now the default.