Figured out a non-nightly method for copy-less fs APIs via explicit "Bound" wrapper
What we want to do:

    trait MemoryHandler {
        type Slice<'a>: AsRef<[u8]>;
        fn get_slice<'a>(&'a self) -> Slice<'a>;
    }

But unfortunately this requires GATs, which is only available on
nightly, with lots of warning signs that it is incomplete.

The workaround here is to require the type implementations all be
"lifetime-less", which requires unsafe code in the implementations.

To make this API not completely unsafe, we can wrap these types with a
wrapper that reintroduces the lifetime requirements. This of course
isn't checked by the compiler, because we're faking everything, but
at least means we can correctly borrow-check dependent code.

    trait MemoryHandler {
        type Slice: AsRef<[u8]>;
        fn get_slice<'a>(&'a self) -> Bound<'a, Slice>;
    }