Added small-object-optimization for IoVecSlice and IoVecSliceMut
With this it's possible to take the full path from WASI's write call
to the filesystem without any memory allocaiton.

I don't think this last bit will have a significant impact, especially
as the read/write block size grows in size, but it's a nice optimization
to have for the by-far most common iovec sizes. We'll see what the
benchmarks say.