runtime-manager: Use a separate directory for the Dockerfile.
The main motivation for this is that the rust-psa-crypto crate
makes symbolic links under target/.../build/, and these cause
nitro-cli to get confused and stop with error E49.

The single file needed by the Dockerfile is hard-linked into
dockerdir/. We trust that nitro-cli is not confused by hard links.
However, see https://xkcd.com/2531/

Also, use "--output-file $@.new", then "mv $@.new $@". This means
that if nitro-cli fails for any reason then a subsequent invocation
of "make" will not think that the target file is up-to-date.