Test if we can use psa_crypto.
The "system call" test_psa_crypto is added, in the same way that
fd_create was added. It takes no arguments and does some psa_crypto
operations: it calls psa_crypto::init then generates a key and
signs using it.

rust-examples/linear-regression is modified so that it calls
that function.