pre-commit: Run rustfmt without "--skip-children --unstable-features".
Now we are using a stable compiler people are likely to have the
stable rustfmt, which does not have "--skip-children". Also we have
applied rustfmt uniformly across the codebase so it should not hurt to
apply rustfmt more broadly.

This patch also adds code to check that rustfmt ran correctly. The
problem with rustfmt failing was previously easy to ignore.