CLI: Removed extra CLI_QUICKSTART, kept superset
This deduplicates the two CLI_QUICKSTARTS we had. There originally was
two due to a request of a quicker quickstart for developers who want to
leverage the prebuilt test-collateral, but it has been clear we can't
maintain two files. So deduplicating to a single quickstart to help
maintenance.

If you want to leverage the test-collateral output, you can also just
skip the steps that rebuilds those files.