CLI: Changed our example to show preview of WebAssembly dissassembly
This accomplishes the original goal of that line, to test that the file
was actually built, but also shows a bit of dissassembly and shows how
one could was wasm-objdump to find more info.

Unfortunately wabt isn't available in any apt-get repository until
20.04, at the time of writing we are working on updating our docker
images to 20.04 in parallel, so for now this will just use the wabt that
lives in the wasm-checker directory.