.github/workflows/main.yml: Remove "PROFILE=dev" from "icecap" test.
It turns out that test-server takes a very long time with PROFILE=dev,
so we remove it, though it would be good to understand why it is so
slow and perhaps put it back one day.

Also double the time-out with VERACRUZ_TEST_TIMEOUT=2400 as some
tests have been seen to pass after a run lasting 35 minutes.