Fixed virtio-console IRQ number quantity for icecap-lkvm configuration
The original "fix" was a mistake, these should have an offset of 0x20
since they are "share processor interrupts" (SPI) which can be handled
by any core. Documentation was unfortunately hard to come by.

Also updated the number of virtio MMIO blocks to 1, since we only have a
single MMIO block in lkvm at the moment.