Replace RUNTIME_ENCLAVE_SPAWN_DELAY with "--write-nl".
In runtime_manager_linux.rs, if "--write-nl" was specified,
write a newline when the server is ready.

In veracruz_server_linux.rs, remove RUNTIME_ENCLAVE_SPAWN_DELAY,
specify "--write-nl", and expect to read a newline from the
runtime managers's stdout.

There is a call to nix::unistd::alarm::set to generate SIGALRM in
30 seconds in case the runtime manager never starts up.