Remove the two-second sleep from veracruz_server_nitro.rs.
The sleep seems to be unnecessary: it happens after the connect (in
NitroEnclave::new in io-utils/src/nitro.rs) and adding a longer sleep
at the start of nitro_main in runtime_manager_nitro.rs does not make
the tests fail.

Also bracket the call to connect with calls to alarm::set and
alarm::cancel so we do not wait for ever for the enclave to start.