Add "std" feature to "serde" dependency of virtio-console-server.
We can see that this is a real requirement by trying to build
virtio-console-server on its own. We have survived without it so
far because virtio-console-server was only ever built in the same
workspace together with runtime-manager, which depends on
veracruz-utils, which depends on mbedtls, which has the "default"
feature, that includes "std", that specifies "serde/std".

It is better to make the requirement explicit rather than depend
on the intricacies of Cargo.