workspaces/nitro-*/Makefile: Use strip -p and recursive make.
The changes in nitro-host/Makefile are unimportant.

In nitro-runtime/Makefile, when building "nitro", we first invoke
"cargo build" and "strip -p" and then use a recursive make invocation
to see whether we need to run "nitro-cli". Note the use of "&:", which
might not matter here, but apparently one should use it for multiple
targets built by a single recipe.

The bug we are fixing here is that without the recursive make
invocation a change to the runtime manager's source does not get
propagated through to the EIF file by a single invocation of "make".
This is likely to be a general problem whenever something depends on
the output of "cargo build".