workspaces/icecap-runtime/Makefile: Remove most of the PHONY targets.
If a target with a recipe depends on a phony target, then the recipe
gets run again and again and again, so don't do that.

Instead we create empty files in t/ that record when a target was
last built.

This prevents some things from getting unnecessarily built twice
during a normal build after a fresh checkout.

Also add the "clean" target.