transport-protocol/build.rs: Put generated file in tmp dir and rename.
This is to avoid unnecesssary rebuilds when several other crates have
a path dependency on transport-protocol. Without this change, one of
the crates getting rebuilt causes the timestamp on the generated file
("transport_protocol.rs") to get updated, which makes the other crates
get rebuilt by "cargo build" next time.

A neater solution might be to put the generated file under OUT_DIR,
but currently that seems to be difficult.