examples/data-generators/*/Makefile: Fix dependencies on directories.
Also various minor simplifications and tidying.

A dependency on a directory causes unnecessary rebuilds as make
looks at the modification time of the directory, which gets updated
as files are created in the directory, so the directory ends up
being newer than some of the target files, forcing a rebuild.
An order-only prerequisite is used to avoid this problem.

List of changes:

- Use an "order-only prerequisite" to create $(TARGET_DIR).
- Remove all uses of .SECONDEXPANSION and notdir.
- Remove all "static pattern rules".
- Use "grouped targets" where appropriate.
- Update sources for external data.
- Add moving-average-*/apple_prices.csv.
- Remove trailing / from value of TARGET_DIR.
- Use "quick-clean: clean".
- Remove trailing spaces.